исправлены ошибки, удалены зависимости, установлен tsconfigpath
