+++ package.json
@@ -7,12 +7,10 @@
     "build": "tsc && vite build"
   },
   "dependencies": {
-    "localforage": "^1.10.0",
-    "match-sorter": "^6.3.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-router-dom": "^6.17.0",
-    "sort-by": "^1.2.0"
+    "vite-tsconfig-paths": "^4.2.1"
   },
   "devDependencies": {
     "@types/node": "^20.8.8",
