+++ src/pages/Auth/Auth.tsx
@@ -13,22 +13,16 @@ export const Auth = () => {
     }
   }, [isLoggedIn]);
 
-  const handleLogin = (username: string) => {
-    authenticate(username);
-  };
-
   const handleChange = (event: ChangeEvent<HTMLInputElement>) => {
     setText(event.target.value);
   };
 
   return (
-    <>
-      <div className="flex-col">
-        <input type="text" placeholder="Введите имя" value={text} className="border p-2" onChange={handleChange} />
-        <button disabled={!text} onClick={() => handleLogin(text)}>
-          login
-        </button>
-      </div>
-    </>
+    <div className="flex-col">
+      <input type="text" placeholder="Введите имя" value={text} className="border p-2" onChange={handleChange} />
+      <button disabled={!text} onClick={() => authenticate(text)}>
+        login
+      </button>
+    </div>
   );
 };
