+++ src/pages/Auth/index.ts
@@ -1,1 +1,1 @@
-export { Auth } from './Auth';
+export * from './Auth';
