+++ src/pages/NotFound/NotFound.tsx
@@ -1,7 +1,3 @@
 export const NotFound = () => {
-  return (
-    <>
-      <div>404</div>
-    </>
-  );
+  return <div>404</div>;
 };
