+++ src/pages/NotFound/index.ts
@@ -1,1 +1,1 @@
-export { NotFound } from './NotFound';
+export * from './NotFound';
