+++ src/pages/Table/Table.tsx
@@ -6,11 +6,9 @@ export const Table = () => {
   const { userName } = useContext(AuthContext);
 
   return (
-    <>
-      <div className="bg-amber-500">
-        <NavBar userName={userName} />
-        <div>Table Component with {userName}</div>
-      </div>
-    </>
+    <div className="bg-amber-500">
+      <NavBar userName={userName} />
+      <div>Table Component with {userName}</div>
+    </div>
   );
 };
