+++ src/pages/Table/index.ts
@@ -1,1 +1,1 @@
-export { Table } from './Table';
+export * from './Table';
