+++ vite.config.ts
@@ -1,13 +1,11 @@
 import * as path from 'path';
 import { defineConfig } from 'vite';
 import react from '@vitejs/plugin-react';
+import tsconfigPaths from 'vite-tsconfig-paths';
 
 export default defineConfig({
   server: {
     port: 8083,
   },
-  resolve: {
-    alias: [{ find: '@', replacement: path.resolve(__dirname, 'src') }],
-  },
-  plugins: [react()],
+  plugins: [react(), tsconfigPaths()],
 });
